+++ app/src/main/java/com/example/chat/application/chatexample/CallActivity.java
@@ -51,10 +51,10 @@ public class CallActivity extends AppCompatActivity implements ChatContract.view
 
     private static final String TAG = "CHAT_SDK_CALL";
     public static final long[] VIB_PATTERN = {0, 1000, 1000};
-    private String TOKEN = BaseApplication.getInstance().getString(R.string.token_ahmadsajadi);
-    private final static String Farhad_TOKEN = BaseApplication.getInstance().getString(R.string.token_ahmadsajadi);
-    private final static String Pooria_TOKEN = BaseApplication.getInstance().getString(R.string.token_ahmadsajadi);
-    private final static String Masoud_TOKEN = BaseApplication.getInstance().getString(R.string.token_ahmadsajadi);
+    private String TOKEN = BaseApplication.getInstance().getString(R.string.Farhad_Kheirkhah);
+    private final static String Farhad_TOKEN = BaseApplication.getInstance().getString(R.string.Farhad_Kheirkhah);
+    private final static String Pooria_TOKEN = BaseApplication.getInstance().getString(R.string.Pooria_Pahlevani);
+    private final static String Masoud_TOKEN = BaseApplication.getInstance().getString(R.string.Masoud_Amjadi);
 
     //INTEGRATION