+++ app/src/main/java/com/example/chat/application/chatexample/ChatActivity.java
@@ -167,7 +167,7 @@ public class ChatActivity extends AppCompatActivity
     //local
 
 //
-    private static String TOKEN = BaseApplication.getInstance().getString(R.string.token_ahmadsajadi);
+    private static String TOKEN = BaseApplication.getInstance().getString(R.string.Farhad_Kheirkhah);
     private static String ssoHost = BaseApplication.getInstance().getString(R.string.integration_ssoHost);
     private static String serverName = BaseApplication.getInstance().getString(R.string.integration_serverName);
 
@@ -2622,7 +2622,7 @@ public class ChatActivity extends AppCompatActivity
     @Override
     protected void onPause() {
         super.onPause();
-        presenter.onStop();
+        presenter.logOut();
     }
 
     @Override