+++ podchat/src/main/java/com/fanap/podchat/chat/Chat.java
@@ -278,11 +278,8 @@ import java.io.BufferedReader;
 import java.io.File;
 import java.io.FileInputStream;
 import java.io.FileNotFoundException;
-import java.io.FileOutputStream;
 import java.io.IOException;
-import java.io.InputStream;
 import java.io.InputStreamReader;
-import java.io.OutputStream;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.Collections;
@@ -294,6 +291,7 @@ import java.util.Map;
 import java.util.Objects;
 import java.util.Set;
 import java.util.UUID;
+import java.util.concurrent.atomic.AtomicBoolean;
 
 import javax.inject.Inject;
 
@@ -431,6 +429,7 @@ public class Chat extends AsyncAdapter {
 
 
     private static PodCallAudioCallServiceManager audioCallManager;
+    private boolean requestToClose;
 
 
     /**
@@ -480,6 +479,7 @@ public class Chat extends AsyncAdapter {
 
             async.rawLog(BuildConfig.DEBUG);
             async.isLoggable(BuildConfig.DEBUG);
+            async.setReconnectOnClose(false);
 
             instance = new Chat();
             gson = new GsonBuilder().setPrettyPrinting().create();
@@ -755,9 +755,16 @@ public class Chat extends AsyncAdapter {
 
             //it listen to turning on and off wifi or mobile data and accessing to internet
 
+            AtomicBoolean initState = new AtomicBoolean(true);
+
             networkStateReceiver.addListener(new NetworkStateListener() {
                 @Override
                 public void networkAvailable() {
+
+                    if(initState.get()){
+                        initState.set(false);
+                        return;
+                    }
 
                     Log.i(TAG, "Network State Changed, Available");
                     tryToConnectOrReconnect();
@@ -767,6 +774,11 @@ public class Chat extends AsyncAdapter {
 
                 @Override
                 public void networkUnavailable() {
+
+                    if(initState.get()){
+                        initState.set(false);
+                        return;
+                    }
 
                     Log.e(TAG, "Network State Changed, Unavailable");
                     closeSocketServer();
@@ -810,11 +822,17 @@ public class Chat extends AsyncAdapter {
 
     public void closeChat() {
         try {
+            requestToClose = true;
             stopTyping();
-            context.unregisterReceiver(networkStateReceiver);
-            if (pinger != null) pinger.stopPing();
-            closeSocketServer();
-            PodNotificationManager.unRegisterReceiver(context);
+//            context.unregisterReceiver(networkStateReceiver);
+            if (pinger != null) {
+                pinger.setRequestToClose(requestToClose);
+            }
+
+            if (chatReady || ASYNC_READY.equals(chatState))
+                async.closeSocket();
+//            closeSocketServer();
+//            PodNotificationManager.unRegisterReceiver(context);
         } catch (Exception ex) {
 
             if (log) {
@@ -822,6 +840,17 @@ public class Chat extends AsyncAdapter {
                 Log.w(TAG, "Pinger has been stopped");
             }
         }
+    }
+
+    public void resumeChat() {
+        if(requestToClose){
+            requestToClose = false;
+            tryToConnectOrReconnect();
+            if (pinger != null) {
+                pinger.setRequestToClose(false);
+            }
+        }
+
     }
 
     private synchronized void closeSocketServer() {
@@ -836,7 +865,7 @@ public class Chat extends AsyncAdapter {
 
             chatState = CLOSED;
 
-            scheduleForReconnect();
+//            scheduleForReconnect();
 
         } catch (Exception e) {
             e.printStackTrace();
@@ -846,15 +875,11 @@ public class Chat extends AsyncAdapter {
     }
 
     private void tryToConnectOrReconnect() {
-
         if (isAsyncReady()) {
-
             pingAfterSetToken();
-
         } else {
             scheduleForReconnect();
         }
-
     }
 
     private boolean isAsyncReady() {
@@ -870,7 +895,10 @@ public class Chat extends AsyncAdapter {
         connectHandler.postDelayed(new Runnable() {
             @Override
             public void run() {
-                if (log) Log.i(TAG, "The Connection Watcher is trying to reconnect...");
+
+
+                if (requestToClose) return;
+
                 if (connectNumberOfRetry < maxReconnectStepTime) {
                     connectNumberOfRetry = connectNumberOfRetry * 2;
                 } else {
@@ -883,9 +911,7 @@ public class Chat extends AsyncAdapter {
                         || shouldReconnectAtOpenState() || shouldReconnectAtConnectingState();
 
                 if (shouldReconnect) {
-
                     async.connect(socketAddress, appId, serverName, token, ssoHost, "");
-
                 } else if (!chatReady) {
                     if (isAsyncReady()) {
                         pingAfterSetToken();