+++ podchat/src/main/java/com/fanap/podchat/util/NetworkUtils/NetworkPingSender.java
@@ -61,6 +61,8 @@ public class NetworkPingSender {
 
     private boolean hasPing = false;
 
+    private boolean requestToClose = false;
+
 
     public void setDisConnectionThreshold(int disConnectionThreshold) {
 
@@ -169,20 +171,29 @@ public class NetworkPingSender {
 
         handlerThread.start();
 
+        Handler pingHandler = new Handler(handlerThread.getLooper());
+
 
         Runnable job = new Runnable() {
             @Override
             public void run() {
 
-                ping();
-                new Handler(handlerThread.getLooper()).postDelayed(this, interval);
+                if(!requestToClose){
+                    ping();
+                }
+                pingHandler.postDelayed(this, interval);
 
             }
         };
+        pingHandler.postDelayed(job, interval);
+    }
 
-        new Handler(handlerThread.getLooper()).postDelayed(job, interval);
+    public boolean isRequestToClose() {
+        return requestToClose;
+    }
 
-
+    public void setRequestToClose(boolean requestToClose) {
+        this.requestToClose = requestToClose;
     }
 
     public void stopPing() {